work of Leo and Gaia --> read description
We created some functions to structure better the data.
It is a little bit complicated and we don't think it's really possible to make the loading completely automatised. In particular, for the healthy sybjects it's kinda annoying. For subjects 1 and 2 there is one of the two structures (NO FLOAT and FLOAT) that does not work, but it's probably a problem of the data itself.

Also, the filtering actually looks strange... It seems we do not really apply it. What do you think?

Other point --> we did the mean over trial for the EMG of each muscle and we normalized by the maximum contraction. Notice that the trials did not have all the same length (so they did not last the same amount of time) and we just cut them at the length of the smallest trial ... Does it make sense for you?

BYEEEEEE <3