Visual detection on Z
Sooooo, the visual detection of the events performed on the small hills (the Z coordinate of the signal of each marker) is done. I stocked everything in a structure containing all the events... maybe it would be nice to create a function to transfer for each trial these events in the original structure (ad it was for find_events() ) or adapt the already existing function to split the fait by make it take into account this ground_truth structure.
I added in the main few lines about the visual detection --> OF COURSE THEY ARE COMMENTED SINCE WE DO NOT WANT TO REPEAT THIS HELL EVERY TIME. However, in the same subsection we load the existing structure with the ground_truth events previously detected.
There is also a function called chek_ground_truth_events(), which plot the stair (Y coordinate) together with the hill (Z coordinate) and the detected points to visualize how well it worked... By looking at the hills it seems really nice and good, but it's true that sometimes on the stair it does not correspond at all to what we found...so maybe it can really improve something? Let's hope :)

Nothing to add,
have fun,
kiss kiss
Gossip Boy