Appending Events
So, I coded two functions really similar to append gait events and something else... the aim was just to stock in the initial structure the events I found with the ground truth and, in particular, to stock them with the same names we used until now --> in order to not have to change everything.

This seems work, so nice :)
I also tried manually to plot some raw signal of the EMG and, indeed, it's true that now the HS is located at the half of the burst of the TA, which makes more sense, because it gives more the habitual pattern of the two peaks with nothing in the middle. However, I just did it for three muscles... so of course we have to work on that.

And so, we can now focus on the EMG and try to solve this problem :)

Bye Bye